+++ lib/audio_tone_method_channel.dart
@@ -46,6 +46,14 @@ class MethodChannelAudioTone extends AudioTonePlatform {
     );
   }
 
+  @override
+  Future<void> setLightFlashingMagnificationFactor(double factor) async {
+    await methodChannel.invokeMethod<void>(
+      'setLightFlashingMagnificationFactor',
+      factor,
+    );
+  }
+
   /// 设置字母之间的间隔时长（点的倍数）
   /// Set Duration of a Letter Interval (Dot counts)
   /// default value 3 dots, range 1-5 dots / 默认3个点，范围1-5个点
@@ -116,7 +124,6 @@ class MethodChannelAudioTone extends AudioTonePlatform {
   /// [onComplete] 播放完成回调 / Playback complete callback
   @override
   StreamSubscription<dynamic> playStream(String morseCode) {
-    _streamSubscription?.cancel();
     _streamSubscription = null;
     _streamSubscription = eventChannel
         .receiveBroadcastStream(morseCode)